% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{add_simulation_id_column}
\alias{add_simulation_id_column}
\title{Add simulation ID column (id's starting at 1 and consecutive). Original column ID
will be replaced by the new simulation ID column, after being renamed into
ORIGINAL_ID column.}
\usage{
add_simulation_id_column(dataset, id = "ID")
}
\arguments{
\item{dataset}{NONMEM dataset}

\item{id}{current identifier column, default is 'ID'}
}
\value{
updated data frame
}
\description{
Add simulation ID column (id's starting at 1 and consecutive). Original column ID
will be replaced by the new simulation ID column, after being renamed into
ORIGINAL_ID column.
}
