% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{append_original_id}
\alias{append_original_id}
\title{Append original ID to simulation output if it exists in the dataset.}
\usage{
append_original_id(x, dataset)
}
\arguments{
\item{x}{Campsis output}

\item{dataset}{engine table OR Campsis dataset}
}
\value{
updated output
}
\description{
Append original ID to simulation output if it exists in the dataset.
}
