% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{appendOriginalId}
\alias{appendOriginalId}
\title{Append original ID to simulation output if it exists in the dataset.}
\usage{
appendOriginalId(x, dataset)
}
\arguments{
\item{x}{CAMPSIS output}

\item{dataset}{engine table OR CAMPSIS dataset}
}
\value{
updated output
}
\description{
Append original ID to simulation output if it exists in the dataset.
}
