% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_core.R
\name{are_equal}
\alias{are_equal}
\title{Are vectors equal given the tolerance.}
\usage{
are_equal(x, xref, tolerance, id, type)
}
\arguments{
\item{x}{first vector}

\item{xref}{second vector}

\item{tolerance}{relative tolerance}

\item{id}{subject ID (for troubleshooting)}

\item{type}{vector type (for troubleshooting)}
}
\value{
logical vector
}
\description{
Are vectors equal given the tolerance.
}
