% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{areEqual}
\alias{areEqual}
\title{Are vectors equal given the tolerance.}
\usage{
areEqual(x, xref, tolerance, id, type)
}
\arguments{
\item{x}{first vector}

\item{xref}{second vector}

\item{tolerance}{relative tolerance}

\item{id}{subject ID (for troubleshooting)}

\item{type}{vector type (for troubleshooting)}
}
\value{
logical vector
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`areEqual()` is deprecated in favor of `are_equal()`.
}
