% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_metrics.R
\name{assessPackage}
\alias{assessPackage}
\title{Access package using the riskmetric package.}
\usage{
assessPackage(package, successful_tests)
}
\arguments{
\item{package}{package name}

\item{successful_tests}{were tests successful?}
}
\value{
human readable table
}
\description{
Access package using the riskmetric package.
}
