% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytex.R
\name{check_tinytex_installation}
\alias{check_tinytex_installation}
\title{Check TinyTeX installation}
\usage{
check_tinytex_installation()
}
\value{
TRUE if TinyTeX is installed correctly, FALSE otherwise
}
\description{
Check TinyTeX installation
}
