% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{checkTinyTEXInstallation}
\alias{checkTinyTEXInstallation}
\title{Check TinyTeX installation.}
\usage{
checkTinyTEXInstallation()
}
\value{
TRUE if TinyTeX is installed correctly, FALSE otherwise
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`checkTinyTEXInstallation()` is deprecated in favor of `check_tinytex_installation()`.
}
