% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{clean_package_results}
\alias{clean_package_results}
\title{Clean package results.}
\usage{
clean_package_results(raw)
}
\arguments{
\item{raw}{raw results}
}
\value{
cleaned results
}
\description{
Clean package results.
}
