% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{cleanPackageResults}
\alias{cleanPackageResults}
\title{Clean package results.}
\usage{
cleanPackageResults(raw)
}
\arguments{
\item{raw}{raw results}
}
\value{
cleaned results
}
\description{
Clean package results.
}
