% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{collect_package_warnings}
\alias{collect_package_warnings}
\title{Collect package warnings}
\usage{
collect_package_warnings(x)
}
\arguments{
\item{x}{package test results}
}
\value{
a tibble with 2 columns: 'warning', 'occurrences'
}
\description{
Collect package warnings
}
