% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_core.R
\name{compare}
\alias{compare}
\title{Compare NONMEM results with CAMPSIS results, according to the given tolerance.}
\usage{
compare(
  ipred,
  campsis,
  variables,
  tolerance,
  dest = "rxode2",
  ipred_source = "NONMEM"
)
}
\arguments{
\item{ipred}{individual predictions to be compared with (=reference results)}

\item{campsis}{Campsis results, data frame}

\item{variables}{variables to be compared}

\item{tolerance}{comparison tolerance}

\item{dest}{destination engine that was used in Campsis}

\item{ipred_source}{source of individual predictions, default is "NONMEM" (informative field)}
}
\value{
logical vector
}
\description{
Compare NONMEM results with CAMPSIS results, according to the given tolerance.
}
