% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_suite_credentials.R
\name{Credentials}
\alias{Credentials}
\title{Qualification suite credentials.}
\usage{
Credentials(cert, key, passphrase)
}
\arguments{
\item{cert}{path to client certificate}

\item{key}{path to private key}

\item{passphrase}{private key passphrase}
}
\value{
a qualification suite credentials object
}
\description{
Qualification suite credentials.
}
