% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{decrypt_file}
\alias{decrypt_file}
\title{Decrypt a file (no prompt).}
\usage{
decrypt_file(file, private_key, passphrase = NULL)
}
\arguments{
\item{file}{path to file to decrypt.}

\item{private_key}{path to private key}

\item{passphrase}{passphrase}
}
\value{
nothing
}
\description{
Decrypt a file (no prompt).
}
