% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{decryptFolder}
\alias{decryptFolder}
\title{Decrypt folder.}
\usage{
decryptFolder(from, to, private_key, passphrase)
}
\arguments{
\item{from}{original folder}

\item{to}{destination folder}

\item{private_key}{path to private key}

\item{passphrase}{passphrase}
}
\value{
nothing
}
\description{
Decrypt folder.
}
