% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{encrypt_file}
\alias{encrypt_file}
\title{Encrypt a file.}
\usage{
encrypt_file(file, public_key)
}
\arguments{
\item{file}{path to file to encrypt.}

\item{public_key}{path to public key}
}
\value{
nothing
}
\description{
Encrypt a file.
}
