% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{encrypt_folder}
\alias{encrypt_folder}
\title{Encrypt folder.}
\usage{
encrypt_folder(from, to, public_key)
}
\arguments{
\item{from}{original folder}

\item{to}{destination folder}

\item{public_key}{path to public key}
}
\value{
nothing
}
\description{
Encrypt folder.
}
