% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{executeNONMEM}
\alias{executeNONMEM}
\title{Execute NONMEM. PsN will be called automatically by R. 
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.}
\usage{
executeNONMEM(outputFolder, reexecuteNONMEM = T)
}
\arguments{
\item{outputFolder}{output folder of prepared NONMEM files}

\item{reexecuteNONMEM}{force re-execute NONMEM if results already exist}
}
\description{
Execute NONMEM. PsN will be called automatically by R. 
Prepared control stream 'model.mod' is executed automatically and NONMEM results
are returned in the form of a data frame.
}
