% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{fix_rxode_bug}
\alias{fix_rxode_bug}
\title{Fix very annoying bug in current version of RxODE (> v1.1.0) or rxode2 (2.0.7).
Description of the bug:
If the model has a lag time and if the dataset does have an observation at time 0,
RxODE still outputs the time 0.}
\usage{
fix_rxode_bug(campsis, model, dataset, dest)
}
\arguments{
\item{campsis}{Campsis output}

\item{model}{Campsis model}

\item{dataset}{engine table OR Campsis dataset}

\item{dest}{destination engine}
}
\value{
the corrected output if the bug was not present if RxODE
}
\description{
Fix very annoying bug in current version of RxODE (> v1.1.0) or rxode2 (2.0.7).
Description of the bug:
If the model has a lag time and if the dataset does have an observation at time 0,
RxODE still outputs the time 0.
}
