% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{fixRxODEBug}
\alias{fixRxODEBug}
\title{Fix very annoying bug in current version of RxODE (> v1.1.0) or rxode2 (2.0.7).
Description of the bug:
If the model has a lag time and if the dataset does have an observation at time 0,
RxODE still outputs the time 0.}
\usage{
fixRxODEBug(campsis, model, dataset, dest)
}
\arguments{
\item{campsis}{CAMPSIS output}

\item{model}{CAMPSIS model}

\item{dataset}{engine table OR CAMPSIS dataset}

\item{dest}{destination engine}
}
\value{
the corrected output if the bug was not present if RxODE
}
\description{
Fix very annoying bug in current version of RxODE (> v1.1.0) or rxode2 (2.0.7).
Description of the bug:
If the model has a lag time and if the dataset does have an observation at time 0,
RxODE still outputs the time 0.
}
