% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{get_certificate_information}
\alias{get_certificate_information}
\title{Get more information on the given certificate.}
\usage{
get_certificate_information(cert)
}
\arguments{
\item{cert}{certificate}
}
\value{
certificate information
}
\description{
Get more information on the given certificate.
}
