% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_summary.R
\name{get_table}
\alias{get_table}
\alias{get_table,qualification_summary,numeric-method}
\title{Get qualification results (table form) for given subject ID.}
\usage{
get_table(summary, id)

\S4method{get_table}{qualification_summary,numeric}(summary, id)
}
\arguments{
\item{summary}{summary object}

\item{id}{subject ID}
}
\value{
a data frame with all variables being compared
}
\description{
Get qualification results (table form) for given subject ID.
}
