% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getCACertificate}
\alias{getCACertificate}
\title{Get the CA certificate.}
\usage{
getCACertificate()
}
\value{
CA certificate, character vector
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getCACertificate()` is deprecated in favor of `get_ca_certificate()`.
}
