% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getCertificateInformation}
\alias{getCertificateInformation}
\title{Get more information on the given certificate.}
\usage{
getCertificateInformation(cert)
}
\arguments{
\item{cert}{certificate}
}
\value{
certificate information
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getCertificateInformation()` is deprecated in favor of `get_certificate_information()`.
}
