% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getPlot}
\alias{getPlot}
\title{Get qualification plot for given subject ID and variable.}
\usage{
getPlot(summary, id, variable)
}
\arguments{
\item{summary}{summary object}

\item{id}{subject ID}

\item{variable}{compared variable}
}
\value{
a plot
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getPlot()` is deprecated in favor of `get_plot()`.
}
