% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_qualification_template}
\alias{model_qualification_template}
\title{Model qualification template.}
\format{
An object of class \code{character} of length 1.
}
\usage{
model_qualification_template
}
\description{
Model qualification template.
}
\keyword{datasets}
