% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_summary.R
\name{passed}
\alias{passed}
\alias{passed,qualification_summary-method}
\title{Say if the qualification passed or not.}
\usage{
passed(summary)

\S4method{passed}{qualification_summary}(summary)
}
\arguments{
\item{summary}{qualification summary object}
}
\value{
TRUE/FALSE
}
\description{
Say if the qualification passed or not.
}
