% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{print_summary_kable}
\alias{print_summary_kable}
\title{Print summary kable.}
\usage{
print_summary_kable(x)
}
\arguments{
\item{x}{data frame}
}
\value{
printed kable
}
\description{
Print summary kable.
}
