% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{printSummaryKable}
\alias{printSummaryKable}
\title{Print summary kable.}
\usage{
printSummaryKable(x)
}
\arguments{
\item{x}{data frame}
}
\value{
printed kable
}
\description{
Print summary kable.
}
