% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{qualification_passed}
\alias{qualification_passed}
\title{Passed method.}
\usage{
qualification_passed(results)
}
\arguments{
\item{results}{results}
}
\value{
TRUE or FALSE
}
\description{
Passed method.
}
