% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_suite.R
\docType{class}
\name{qualification_suite-class}
\alias{qualification_suite-class}
\title{Qualification suite class.}
\description{
Qualification suite class.
}
\section{Slots}{

\describe{
\item{\code{path}}{path to qualification suite (zip file)}

\item{\code{credentials}}{qualification suite credentials}

\item{\code{nmodels}}{number of models to qualify}
}}

