% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_suite_credentials.R
\docType{class}
\name{qualification_suite_credentials-class}
\alias{qualification_suite_credentials-class}
\title{Qualification suite credentials class.}
\description{
Qualification suite credentials class.
}
\section{Slots}{

\describe{
\item{\code{certificate_path}}{path to client certificate}

\item{\code{certificate}}{client certificate, PEM format}

\item{\code{private_key_path}}{path to private key}

\item{\code{passphrase}}{private key passphrase}
}}

