% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_summary.R
\docType{class}
\name{qualification_summary-class}
\alias{qualification_summary-class}
\title{Qualification summary class.}
\description{
Qualification summary class.
}
\section{Slots}{

\describe{
\item{\code{ids}}{list of subject ID's being qualified}

\item{\code{original_ids}}{list of original ID's}

\item{\code{variables}}{variables from dataset that were qualified}

\item{\code{summary}}{qualification summary, data frame}

\item{\code{plots}}{list of qualification plots}

\item{\code{tables}}{list of qualification tables}

\item{\code{ipred_source}}{source of individual predictions, e.g. "NONMEM"}

\item{\code{dest}}{destination engine, e.g. "rxode2"}

\item{\code{model_name}}{name of the model being qualified, e.g. "RUN001"}

\item{\code{tolerance}}{relative tolerance used for qualification, numeric}
}}

