% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{qualificationPassed}
\alias{qualificationPassed}
\title{Passed method.}
\usage{
qualificationPassed(results)
}
\arguments{
\item{results}{results}
}
\value{
TRUE or FALSE
}
\description{
Passed method.
}
