% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_suite.R
\name{QualificationSuite}
\alias{QualificationSuite}
\title{Qualification suite.}
\usage{
QualificationSuite(path, credentials, nmodels = 97L)
}
\arguments{
\item{path}{path to qualification suite (zip file)}

\item{credentials}{qualification suite credentials, see ?Credentials}

\item{nmodels}{number of models to qualify, max number is 97. Please note that the models will be drawn at random.}
}
\value{
a qualification suite object
}
\description{
Qualification suite.
}
