% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{qualify}
\alias{qualify}
\title{Qualify function.}
\usage{
qualify(
  x,
  model,
  dataset,
  variables,
  tolerance = 0.01,
  outputFolder,
  reexecuteNONMEM = T,
  compartments = NULL,
  dest = "rxode2",
  seed = 1,
  settings = Settings(NOCB(TRUE))
)
}
\arguments{
\item{x}{campsistrans object, if NULL, NONMEM files must exist}

\item{model}{CAMPSIS model}

\item{dataset}{CAMPSIS dataset}

\item{variables}{variables to compare}

\item{tolerance}{relative tolerance accepted when comparing 2 values}

\item{outputFolder}{output folder for qualification}

\item{reexecuteNONMEM}{force re-execute NONMEM if results already exist}

\item{compartments}{compartment indexes to output, numeric vector}

\item{dest}{destination engine to use}

\item{seed}{simulation/table export seed, default value is 1}

\item{settings}{simulation settings, please note NOCB is enabled by default}
}
\value{
qualification summary
}
\description{
Qualify function.
}
