% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{qualify_delegate}
\alias{qualify_delegate}
\title{Qualify function.}
\usage{
qualify_delegate(
  x,
  model,
  engineTable,
  nmTable,
  variables,
  tolerance,
  outputFolder,
  reexecuteNONMEM = T,
  compartments = NULL,
  dest,
  seed,
  settings
)
}
\arguments{
\item{x}{campsistrans object, if NULL, NONMEM files must exist}

\item{model}{CAMPSIS model}

\item{engineTable}{engine table OR CAMPSIS dataset}

\item{nmTable}{NONMEM dataset, data frame form}

\item{variables}{variables to compare}

\item{tolerance}{relative tolerance accepted when comparing 2 values}

\item{outputFolder}{output folder for qualification}

\item{reexecuteNONMEM}{force re-execute NONMEM if results already exist}

\item{compartments}{compartment indexes to output, numeric vector}

\item{dest}{destination engine to use}

\item{seed}{simulation/table export seed, default value is 1}

\item{settings}{simulation settings}
}
\value{
qualification summary
}
\description{
Qualify function.
}
