% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_results_parser.R
\name{read.nonmem}
\alias{read.nonmem}
\title{Read all sorts of NONMEM tables (extensions ext, tab, cov, etc, ...).}
\usage{
read.nonmem(file)
}
\arguments{
\item{file}{the file to be read}
}
\value{
a list of data frame (one dataframe per table)
}
\description{
Read all sorts of NONMEM tables (extensions ext, tab, cov, etc, ...).
}
