% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_results_parser.R
\name{read.nonmemtable}
\alias{read.nonmemtable}
\title{Read a NONMEM table.}
\usage{
read.nonmemtable(content)
}
\arguments{
\item{content}{table content}
}
\value{
a dataframe
}
\description{
Read a NONMEM table.
}
