% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{renderReport}
\alias{renderReport}
\title{Render report with rmarkdown.}
\usage{
renderReport(
  results,
  packages,
  fullname,
  initials = NULL,
  output_dir,
  qualification_suite
)
}
\arguments{
\item{results}{results}

\item{packages}{packages}

\item{fullname}{fullname}

\item{initials}{username initials to be used in the filename. If not provided, initials are deduced from the fullname.}

\item{output_dir}{output directory}

\item{qualification_suite}{qualification suite object}
}
\value{
nothing
}
\description{
Render report with rmarkdown.
}
