% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{runQualification}
\alias{runQualification}
\title{Get OS name.}
\usage{
runQualification(
  packages,
  fullname,
  initials = NULL,
  output_dir = getwd(),
  qualification_suite = NULL,
  cpu = 6L,
  skip_vdiffr = TRUE,
  skip_python = TRUE
)
}
\arguments{
\item{packages}{packages to qualify}

\item{fullname}{fullname of the person running the qualification}

\item{initials}{username initials to be used in the filename. If not provided, initials are deduced from the fullname.}

\item{output_dir}{output directory}

\item{qualification_suite}{qualification suite object, type ?QualificationSuite for more information}

\item{cpu}{number of workers to be used, default is 6 workers}

\item{skip_vdiffr}{skip Vdiffr tests, default is TRUE}

\item{skip_python}{skip Python tests (e.g. Pharmpy in campsistrans), default is TRUE}
}
\value{
the OS name
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`runQualification()` is deprecated in favor of `run_qualification()`.
}
