% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{runQualificationCore}
\alias{runQualificationCore}
\title{Run qualification core.}
\usage{
runQualificationCore(packages, qualification_suite = NULL, cpu = 6)
}
\arguments{
\item{packages}{packages to qualify}

\item{qualification_suite}{qualification suite object}

\item{cpu}{number of workers to be used}
}
\value{
summarised results
}
\description{
Run qualification core.
}
