% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{set_default_obs_cmt}
\alias{set_default_obs_cmt}
\alias{set_default_obs_cmt,dataset,integer-method}
\alias{set_default_obs_cmt,arm,integer-method}
\alias{set_default_obs_cmt,observations,integer-method}
\title{Set the default observed compartment index.}
\usage{
set_default_obs_cmt(object, index)

\S4method{set_default_obs_cmt}{dataset,integer}(object, index)

\S4method{set_default_obs_cmt}{arm,integer}(object, index)

\S4method{set_default_obs_cmt}{observations,integer}(object, index)
}
\arguments{
\item{object}{generic object}

\item{index}{index of the observed compartment}
}
\value{
updated object
}
\description{
Set the default observed compartment index.
}
