% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{setDefaultObsCmt}
\alias{setDefaultObsCmt}
\alias{setDefaultObsCmt,dataset,integer-method}
\alias{setDefaultObsCmt,arm,integer-method}
\alias{setDefaultObsCmt,observations,integer-method}
\title{Set the default observed compartment index.}
\usage{
setDefaultObsCmt(object, index)

\S4method{setDefaultObsCmt}{dataset,integer}(object, index)

\S4method{setDefaultObsCmt}{arm,integer}(object, index)

\S4method{setDefaultObsCmt}{observations,integer}(object, index)
}
\arguments{
\item{object}{generic object}

\item{index}{index of the observed compartment}
}
\value{
updated object
}
\description{
Set the default observed compartment index.
}
