% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise results.}
\usage{
summarise_results(results)
}
\arguments{
\item{results}{results}
}
\value{
summarised results
}
\description{
Summarise results.
}
