% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{summariseResults}
\alias{summariseResults}
\title{Summarise results.}
\usage{
summariseResults(results)
}
\arguments{
\item{results}{results}
}
\value{
summarised results
}
\description{
Summarise results.
}
