% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_summary.R
\name{write,qualification_summary,character-method}
\alias{write,qualification_summary,character-method}
\title{Export qualification summary to PDF file.}
\usage{
\S4method{write}{qualification_summary,character}(
  object,
  file,
  original_model = NULL,
  failed_only = TRUE,
  debug_tables = TRUE,
  notes = NULL,
  ...
)
}
\arguments{
\item{object}{qualification summary object}

\item{file}{output file name}

\item{original_model}{original model that was imported (e.g. RUN001.ctl), character, default is NULL}

\item{failed_only}{if TRUE, only failed subjects will be included in the report, default is TRUE}

\item{debug_tables}{if TRUE, include debug tables in the report with detailed results (only for failing subjects), default is TRUE}

\item{notes}{additional notes to be included in the report, character, default is NULL}

\item{...}{additional parameters, unused}
}
\description{
Export qualification summary to PDF file.
}
