% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{write_category_results}
\alias{write_category_results}
\title{Write category results.}
\usage{
write_category_results(category, tests, file, ...)
}
\arguments{
\item{category}{category}

\item{tests}{tests}

\item{file}{file}

\item{...}{extra arguments}
}
\value{
nothing
}
\description{
Write category results.
}
