% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{write_file_results}
\alias{write_file_results}
\title{Write file results.}
\usage{
write_file_results(file, categories, ...)
}
\arguments{
\item{file}{file}

\item{categories}{categories}

\item{...}{extra arguments}
}
\value{
nothing
}
\description{
Write file results.
}
