% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{writeCategoryResults}
\alias{writeCategoryResults}
\title{Write category results.}
\usage{
writeCategoryResults(category, tests, file, ...)
}
\arguments{
\item{category}{category}

\item{tests}{tests}

\item{file}{file}

\item{...}{extra arguments}
}
\value{
nothing
}
\description{
Write category results.
}
