% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{writeFileResults}
\alias{writeFileResults}
\title{Write file results.}
\usage{
writeFileResults(file, categories, ...)
}
\arguments{
\item{file}{file}

\item{categories}{categories}

\item{...}{extra arguments}
}
\value{
nothing
}
\description{
Write file results.
}
