% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytex.R
\name{checkTinyTEXInstallation}
\alias{checkTinyTEXInstallation}
\title{Check TinyTeX installation}
\usage{
checkTinyTEXInstallation()
}
\value{
TRUE if TinyTeX is installed correctly, FALSE otherwise
}
\description{
Check TinyTeX installation
}
