% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_core.R
\name{compare}
\alias{compare}
\title{Compare NONMEM results with CAMPSIS results, according to the given tolerance.}
\usage{
compare(nonmem, campsis, variables, tolerance, dest = "RxODE")
}
\arguments{
\item{nonmem}{NONMEM results, dataframe}

\item{campsis}{CAMPSIS results, dataframe}

\item{variables}{variables to be compared}

\item{tolerance}{comparison tolerance}

\item{dest}{destination engine that was used in CAMPSIS}
}
\value{
logical vector
}
\description{
Compare NONMEM results with CAMPSIS results, according to the given tolerance.
}
