% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pki.R
\name{getCertificateInformation}
\alias{getCertificateInformation}
\title{Get more information on the given certificate.}
\usage{
getCertificateInformation(cert)
}
\arguments{
\item{cert}{certificate}
}
\value{
certificate information
}
\description{
Get more information on the given certificate.
}
