% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getOSName}
\alias{getOSName}
\title{Get OS name.}
\usage{
getOSName(short = FALSE)
}
\arguments{
\item{short}{if TRUE, returns a short version of the OS name (e.g., "Win11" for Windows 11)}
}
\value{
the OS name
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getOSName()` is deprecated in favor of `get_os_name()`.
}
