% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{getOSName}
\alias{getOSName}
\title{Get OS name.}
\usage{
getOSName(short = FALSE)
}
\arguments{
\item{short}{if TRUE, returns a short version of the OS name (e.g., "Win10" for Windows 10)}
}
\value{
the OS name
}
\description{
Get OS name.
}
