% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_summary.R
\name{getPlot}
\alias{getPlot}
\alias{getPlot,qualification_summary,numeric,character-method}
\title{Get qualification plot for given subject ID and variable.}
\usage{
getPlot(summary, id, variable)

\S4method{getPlot}{qualification_summary,numeric,character}(summary, id, variable)
}
\arguments{
\item{summary}{summary object}

\item{id}{subject ID}

\item{variable}{compared variable}
}
\value{
a plot
}
\description{
Get qualification plot for given subject ID and variable.
}
