% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{getTable}
\alias{getTable}
\title{Get qualification plot for given subject ID and variable.}
\usage{
getTable(summary, id, variable)
}
\arguments{
\item{summary}{summary object}

\item{id}{subject ID}
}
\value{
a data frame with all variables being compared
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`getTable()` is deprecated in favor of `get_table()`.
}
