% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{qualify}
\alias{qualify}
\title{Qualify function.}
\usage{
qualify(
  model,
  dataset,
  ipred,
  variables,
  tolerance = 0.01,
  dest = "rxode2",
  seed = 1,
  settings = Settings(NOCB(TRUE)),
  idref = "ipred"
)
}
\arguments{
\item{model}{Campsis model to be qualified}

\item{dataset}{Campsis dataset or data frame to be qualified}

\item{ipred}{individual predictions to be compared with (=reference results)}

\item{variables}{variables to compare}

\item{tolerance}{relative tolerance accepted when comparing 2 values}

\item{dest}{destination engine to use}

\item{seed}{simulation/table export seed, default value is 1}

\item{settings}{simulation settings, please note NOCB is enabled by default}

\item{idref}{ID reference, "dataset" or "ipred" (other IDs will be ignored)}
}
\value{
qualification summary
}
\description{
Qualify function.
}
