% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation_qualification.R
\name{runQualification}
\alias{runQualification}
\title{Run qualification.}
\usage{
runQualification(
  packages,
  fullname,
  initials = NULL,
  output_dir = getwd(),
  qualification_suite = NULL
)
}
\arguments{
\item{packages}{packages to qualify}

\item{fullname}{fullname of the person running the qualification}

\item{initials}{username initials to be used in the filename. If not provided, initials are deduced from the fullname.}

\item{output_dir}{output directory}

\item{qualification_suite}{qualification suite object, type ?QualificationSuite for more information}
}
\value{
TRUE if the qualification passes, FALSE otherwise
}
\description{
Run qualification.
}
